pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wandering-rag"
version = "0.1.0"
description = "Wandering RAG CLI tool"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
dependencies = [
    "click>=8.0.0",
    "qdrant-client>=1.7.0",
    "python-dotenv>=1.0.0",
    "fastembed>=0.6.0",
    "sentence-transformers>=2.5.0",
    "python-frontmatter"
]

[project.scripts]
wandering-rag = "wandering_rag.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/wandering_rag"]
